+++ deltafi-ui/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added import/export Metadata to/from Upload page (#190)
 - Refactored file upload buttons on Delete Policies and Ingress Routing (#239)
 
+### Fixed
+- Fixed bug with batching when resuming < 500 errors (#254)
+- Fixed bug with selecting on By Message error summary (#257)
+
 ## [0.98.2]
 
 ### Fixed